Added a way to translate stuff
You can add translations by adding a file in /localisation/. It should be a header file. There are already 2 files there that are for demonstration purposes. (RIGHT NOW THERE IS NO WAY FOR THE USER TO SWITCH LANGUAGES. ONLY YOU CAN CHANGE IT IN THE CODE. SEE KERNEL.C FOR MORE INFO.)